requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
autopep8==2.0.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
clarifai==9.8.1
clarifai-grpc==9.11.5
colorama==0.4.6
comm==0.2.2
configparser==6.0.0
crispy-bootstrap5==0.7
cryptography==41.0.6
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
Django==5.2.1
django-allauth==0.57.2
django-allauth-2fa==0.12.0
django-crispy-forms==2.0
django-extensions==3.2.3
django-filter==23.2
django-otp==1.6.0
django-recaptcha==4.1.0
django-taggit==4.0.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
EasyProcess==1.1
emoji==2.9.0
exceptiongroup==1.2.2
executing==2.1.0
future==0.18.3
googleapis-common-protos==1.62.0
grpcio==1.60.0
h11==0.14.0
humanize==4.9.0
idna==3.4
importlib-metadata==6.8.0
instapy==0.1.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
Markdown==3.4.3
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
MeaningCloud-python==2.0.0
nest-asyncio==1.6.0
numpy==1.26.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
parso==0.8.4
Pillow==10.1.0
platformdirs==4.3.6
plyer==2.1.0
prompt_toolkit==3.0.48
protobuf==4.25.1
psutil==6.0.0
pure_eval==0.2.3
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.18.0
PyJWT==2.7.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-rapidjson==1.14
python3-openid==3.2.0
pytz==2023.3
PyVirtualDisplay==3.0
pywin32==306
PyYAML==3.13
pyzmq==26.2.0
qrcode==8.2
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.16.2
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
tornado==6.4.1
tqdm==4.64.1
traitlets==5.14.3
trio==0.23.2
trio-websocket==0.11.1
tritonclient==2.34.0
typing_extensions==4.12.2
tzdata==2022.6
urllib3==2.0.7
wcwidth==0.2.13
wsproto==1.2.0
zipp==3.15.0
zxcvbn-python==4.4.24
